Revert "docs: use standard header on v4 page"
This reverts commit 403c681fd08ff386b24555b4f2cadcdcf9813c62.

While the previous commit worked, it prevents clients from seeing the
full URL of links on the page. This has been an issue with v1-v3 but
I had assumed the issue was with the preview version of the API blueprint
docs. I'll need to investigate to see if it's really an issue with the
site. If so I may remove the header from all pages.