spec: correct urls for Get Samples
The two operations

- Get Samples for a Network Device Attribute
- Get Samples for an Object Attribute

had incorrect URLs specified. The URLs that were specified were
the ones that were intended but the implementation didn't catch up
to the spec until v4. So this fixes the URLs to match the
implementation. Specifically the word `attributes` is used in the
URL instead of `trendedAttributes`.

/networkDevices/id/trendedAttributes/attributeId/samples ->
  /networkDevices/id/attributes/attributeId/samples

/objects/id/trendedAttributes/attributeId/samples ->
  /objects/id/attributes/attributeId/samples

This represents the changes from
d09933c05c9df5d6fbbf4f8588be342e85ca76a7

https://github.com/jci-products/cp-metasys-rest-api-v1-v3/commit/d09933c05c9df5d6fbbf4f8588be342e85ca76a7