fix redirect for v4 api
Most of the pages when rendered are in their own folder in a
page labeled index.html

So the redirect takes a request for `/v3` and turns it into
`/v3/index.html`. But for whatever reason `v4` just gets copied
over as v4.html. So the redirect just needs to tack on .html