feat: post updates to the PR during and after build
The PR will be updated with a comment
that says Circle CI Preview being generated with a link to the build.
When the preview is available that comment will be updated with a link
to the preview.

If the PR is updated later, the same process is repeated. First
the comment is changed to say preview is being generated; and then
it's updated to point to the preview.