add 14.1 documentation
This new document (v6-14-1) is a copy of v6.html with some
manual modifications to the intro to tailor it to 14.1. As there
are very few changes to the api from 14 to 14.1 this document represents
the majority of the 14.1 spec. However, in the near future the
actual v6 spec for 14.1 will be uploaded.